refactor: NotificationManager 책임 분리
1. NotificationPermissionManager - 알림 권한 관리
2. FCMTokenManager - FCM 토큰 및 서버 통신
3. NotificationParser - 푸시 데이터 파싱
4. ChatNotificationHandler - 채팅 알림 처리 로직
5. BadgeCountManager - 앱 배지 카운트 관리
6. NotificationNavigator - 알림 탭 네비게이션
